requirements.txt
PySide6>=6.6
opencv-python-headless>=4.9
onvif-zeep>=0.2.12
wsdiscovery>=2.0.0
keyring>=25.0
Pillow>=10.0
httpx>=0.27.0
psutil>=5.9.0
